Noise vector field: continuous (fractional) octave count
octaves becomes a float on both the CPU (Noise.Sum / NoiseSamplerProperties)
and GPU (NoiseVectorField.compute) paths: the integer part is summed at full
amplitude and the fractional part fades in the next octave, so detail ramps
in smoothly instead of popping as the slider crosses an integer. The
component now sets numOctaves as a float to match.

+++ Assets/Vector Fields/Vector Field/Components/NoiseVectorFieldComponent.cs
@@ -58,7 +58,7 @@ public class NoiseVectorFieldComponent : VectorFieldComponent {
         NoiseVectorFieldComputeShader.SetFloat("frequency", noiseSampler.properties.frequency);
         NoiseVectorFieldComputeShader.SetFloat("persistence", noiseSampler.properties.persistence);
         NoiseVectorFieldComputeShader.SetFloat("lacunarity", noiseSampler.properties.lacunarity);
-        NoiseVectorFieldComputeShader.SetInt("numOctaves", noiseSampler.properties.octaves);
+        NoiseVectorFieldComputeShader.SetFloat("numOctaves", noiseSampler.properties.octaves);
         NoiseVectorFieldComputeShader.SetFloat("vortexAngle", vortexAngle);
 
         // Calculate the number of thread groups
